Fixed death bug
Crashing if an character was killed on the frame they activated an ability. Well, a more detailed explanation is that if an ability coudl self-harm, then it could apply it's effect to the caster on activate ability, essentially leading it to killing the caster, all before it the program could exit InternalActivateABility scope lock, which meant it tried to destroy the actor when abilitylistscopelockcount was greater than 0. Crashing the game.